src/requirements.txt
anthropic==0.30.0
python-dotenv==1.0.1
multion==1.3.4
pyautogui==0.9.54
streamlit==1.36.0
requests